Fix jumping
The normal to the surface is checked and if the angle between it and (0, 1, 0) is less than slope limit of the character controller jumping becomes possible. To prevent getting stuck on the surfaces where you can't jump sliding was implemented in the form of moving in the direction of hit normal.